backend/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycryptodome==3.23.0
pymongo==4.15.3
python-dotenv==1.1.1
Werkzeug==3.1.3
flask-socketio==5.3.6
python-socketio==5.11.1
requests==2.31.0
beautifulsoup4==4.12.3
flask-jwt-extended==4.6.0
gunicorn==22.0.0
redis==5.0.4
flask-limiter
